feat(UGC-7372) - Рефакторинг: Декомпозиция кода
